refactor(ui-layer.service.ts): Moved the ui layer service to the services directory
