+++ src/BurstChat.Client/src/app/components/chat/chat-channel/chat-channel.component.ts
@@ -5,7 +5,7 @@ import { Message } from 'src/app/models/chat/message';
 import { ChannelConnectionOptions } from 'src/app/models/chat/channel-connection-options';
 import { NotifyService } from 'src/app/services/notify/notify.service';
 import { ChatService } from 'src/app/services/chat/chat.service';
-import { UiLayerService } from 'src/app/modules/chat/services/ui-layer/ui-layer.service';
+import { UiLayerService } from 'src/app/services/ui-layer/ui-layer.service';
 import { ChatMessagesComponent } from 'src/app/components/chat/chat-messages/chat-messages.component';
 import { ChatInfoComponent } from 'src/app/components/chat/chat-info/chat-info.component';
 import { ChatInputComponent } from 'src/app/components/chat/chat-input/chat-input.component';
