+++ src/BurstChat.Client/src/app/components/chat/chat-info/chat-info.component.ts
@@ -10,7 +10,7 @@ import { User } from 'src/app/models/user/user';
 import { RTCSessionContainer } from 'src/app/models/chat/rtc-session-container';
 import { UserService } from 'src/app/services/user/user.service';
 import { RtcSessionService } from 'src/app/services/rtc-session/rtc-session.service';
-import { UiLayerService } from 'src/app/modules/chat/services/ui-layer/ui-layer.service';
+import { UiLayerService } from 'src/app/services/ui-layer/ui-layer.service';
 
 import {
     faCommentAlt,
