+++ src/BurstChat.Client/src/app/components/chat/chat-messages/chat-messages.component.ts
@@ -7,7 +7,7 @@ import { Payload } from 'src/app/models/signal/payload';
 import { ChatConnectionOptions } from 'src/app/models/chat/chat-connection-options';
 import { ChatService } from 'src/app/services/chat/chat.service';
 import { NotifyService } from 'src/app/services/notify/notify.service';
-import { UiLayerService } from 'src/app/modules/chat/services/ui-layer/ui-layer.service';
+import { UiLayerService } from 'src/app/services/ui-layer/ui-layer.service';
 import { ChatMessageComponent } from 'src/app/components/chat/chat-message/chat-message.component';
 
 /**
